move BinaryClassificationMetrics one level up
